Test MapClaims.Valid in isolation
